-  Merge some changes from tcpdump 3.4    -a flag; attempt to convert network and broadcast addresses to names    Improved signal handling    Miscellaneous fixes and typos    OSPF MD5 authentication support
-  -X flag; emacs-hexl print (including ascii)

-  Add ECN bits to TCP and IP headers

-  IKE & IPsec (ESP & AH) support

OK deraadt@

+++ src/usr.sbin/tcpdump/print-isoclns.c
@@ -293,7 +293,7 @@ osi_cksum(register const u_char *p, register u_int len,
 		return 0;
 
 	off[0] = off[1] = 0;
-	while (--len >= 0) {
+	while ((int)--len >= 0) {
 		c0 += *p++;
 		c1 += c0;
 		c0 %= 255;
